submitting homework
The harder this class gets and the more lost I feel, the better I'm feeling about how much I'm learning and the value of the resources being passed along to me. Thanks to Brady and the TAs for the patience and time spent prepping and grading homework, especially the kind I turn in. Have a great weekend.